VenusToken
VENUS TOKEN is a digital asset running on the pre-existing blockchain of Ethereum. In addition to VENUS TOKEN, many different types of tokens run on the Ethereum platform. The VENUS TOKEN has been developed on the Ethereum Network and is classified as a NFT, Non-Fungible ERC721 Token. This standard defines a smart contract framework to allow for the creation, management and trading of our VENUS TOKENS.